requirements.txt
opencv-python>=4.7.0
numpy>=1.24.0
paddleocr>=2.6.0
requests>=2.28.0
pillow>=9.4.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
openai>=0.27.0
websockets>=10.4
